+++ L2J_Mobius_Essence_4.0_DwellingOfSpirits/java/org/l2jmobius/gameserver/model/actor/stat/NpcStat.java
@@ -26,7 +26,7 @@ public class NpcStat extends CreatureStat
 	}
 	
 	@Override
-	public byte getLevel()
+	public int getLevel()
 	{
 		return getActiveChar().getTemplate().getLevel();
 	}
