+++ L2J_Mobius_Essence_4.0_DwellingOfSpirits/java/org/l2jmobius/gameserver/model/actor/stat/StaticObjectStat.java
@@ -32,7 +32,7 @@ public class StaticObjectStat extends CreatureStat
 	}
 	
 	@Override
-	public byte getLevel()
+	public int getLevel()
 	{
 		return (byte) getActiveChar().getLevel();
 	}
